+++ packages/bsky/src/image/uri.ts
@@ -1,4 +1,4 @@
-import { FeatureGatesClient } from '../feature-gates'
+import { UriString, isUriString } from '@atproto/lex'
 import { Options } from './util'
 
 // @NOTE if there are any additions here, ensure to include them on ImageUriBuilder.presets
@@ -11,7 +11,14 @@ export type ImagePreset =
 const PATH_REGEX = /^\/(.+?)\/plain\/(.+?)\/(.+?)(?:@(.+?))?$/
 
 export class ImageUriBuilder {
-  constructor(public endpoint: string) {}
+  public endpoint: UriString
+
+  constructor(endpoint: string) {
+    if (!isUriString(endpoint)) {
+      throw new Error('ImageUriBuilder endpoint must be a valid UriString')
+    }
+    this.endpoint = endpoint
+  }
 
   static presets: ImagePreset[] = [
     'avatar',
@@ -20,52 +27,23 @@ export class ImageUriBuilder {
     'feed_fullsize',
   ]
 
-  getPresetUri(
-    id: ImagePreset,
-    did: string,
-    cid: string,
-    featureGatesClient?: FeatureGatesClient,
-  ): string {
+  getPresetUri(id: ImagePreset, did: string, cid: string): UriString {
     const options = presets[id]
     if (!options) {
       throw new Error(`Unrecognized requested common uri type: ${id}`)
     }
 
-    // TODO: Remove after image migration. It is not ideal to check feature gates outside of handlers with the current setup..
-    const map = featureGatesClient?.checkGates(
-      ['image:remove_format_from_url'],
-      {
-        // This is a workaround. We're trying to use the image owner's DID as if it were the viewer,
-        // while in reality it is the owner of the subject (image) being viewed.
-        // My expectation is that it does the effect of, instead of rolling out gradually by the image viewers,
-        // that it rolls out gradually by the image owners.
-        viewer: did,
-      },
-    )
-    const removeFormat = map?.get('image:remove_format_from_url') ?? false
-    const includeFormat = !removeFormat
+    const path = ImageUriBuilder.getPath({
+      preset: id,
+      did,
+      cid,
+    })
 
-    return (
-      this.endpoint +
-      ImageUriBuilder.getPath({
-        preset: id,
-        did,
-        cid,
-        includeFormat,
-      })
-    )
+    return `${this.endpoint}${path}`
   }
 
-  static getPath(
-    opts: { preset: ImagePreset } & BlobLocation & {
-        includeFormat?: boolean
-      },
-  ) {
-    const { format } = presets[opts.preset]
-    return (
-      `/${opts.preset}/plain/${opts.did}/${opts.cid}` +
-      (opts.includeFormat ? `@${format}` : '')
-    )
+  static getPath(opts: { preset: ImagePreset } & BlobLocation) {
+    return `/${opts.preset}/plain/${opts.did}/${opts.cid}`
   }
 
   static getOptions(
@@ -102,30 +80,32 @@ type BlobLocation = { cid: string; did: string }
 
 export class BadPathError extends Error {}
 
+// @NOTE these prefix settings don't get used anywhere in this package,
+// but they serve as soft documentation of the behavior in production.
 export const presets: Record<ImagePreset, Options> = {
   avatar: {
-    format: 'jpeg', // @TODO switch these formats to webp after rollout
+    format: 'webp',
     fit: 'cover',
     height: 1000,
     width: 1000,
     min: true,
   },
   banner: {
-    format: 'jpeg',
+    format: 'webp',
     fit: 'cover',
     height: 1000,
     width: 3000,
     min: true,
   },
   feed_thumbnail: {
-    format: 'jpeg',
+    format: 'webp',
     fit: 'inside',
     height: 2000,
     width: 2000,
     min: true,
   },
   feed_fullsize: {
-    format: 'jpeg',
+    format: 'webp',
     fit: 'inside',
     height: 1000,
     width: 1000,