+++ packages/lex/lex-schema/src/schema/array.test.ts
@@ -81,7 +81,7 @@ describe('ArraySchema', () => {
         success: false,
         reason: expect.objectContaining({
           message: expect.stringContaining(
-            'Expected array value type (got integer) at $',
+            'Expected array value type (got 3) at $',
           ),
         }),
       })