+++ packages/oauth/oauth-provider-ui/src/components/forms/form-card-async.tsx
@@ -3,27 +3,25 @@ import { FormEvent, ReactNode, useCallback } from 'react'
 import {
   UseAsyncActionOptions,
   useAsyncAction,
-} from '../../hooks/use-async-action.ts'
-import { Override } from '../../lib/util.ts'
+} from '#/hooks/use-async-action.ts'
+import { Override } from '#/lib/util.ts'
 import { ErrorCard } from '../utils/error-card.tsx'
 import { Button } from './button.tsx'
 import { FormCard, FormCardProps } from './form-card.tsx'
 
-export type { AsyncActionController } from '../../hooks/use-async-action.ts'
+export type { AsyncActionController } from '#/hooks/use-async-action.ts'
 
 export type ErrorRender = (data: { error: Error }) => ReactNode
-export const errorRenderDefault: ErrorRender = ({ error }) => (
-  <ErrorCard error={error} />
-)
 
 export type FormCardAsyncProps = Override<
   Override<
-    Omit<FormCardProps, 'cancel' | 'actions' | 'prepend'>,
+    Omit<FormCardProps, 'cancel'>,
     Pick<UseAsyncActionOptions, 'ref' | 'onLoading' | 'onError'>
   >,
   {
     invalid?: boolean
     disabled?: boolean
+    append?: ReactNode
 
     onSubmit: (signal: AbortSignal) => void | PromiseLike<void>
     submitLabel?: ReactNode
@@ -40,12 +38,12 @@ export function FormCardAsync({
   disabled,
 
   onSubmit,
-  submitLabel,
+  submitLabel = <Trans>Submit</Trans>,
 
   onCancel = undefined,
-  cancelLabel,
+  cancelLabel = <Trans>Cancel</Trans>,
 
-  errorRender = errorRenderDefault,
+  errorRender = ({ error }) => <ErrorCard error={error} />,
 
   // UseAsyncActionOptions
   ref,
@@ -54,6 +52,8 @@ export function FormCardAsync({
 
   // FormCardProps
   children,
+  actions,
+  append,
   ...props
 }: FormCardAsyncProps) {
   const { run, loading, error } = useAsyncAction(onSubmit, {
@@ -78,23 +78,20 @@ export function FormCardAsync({
       {...props}
       onSubmit={doSubmit}
       disabled={disabled || loading}
-      prepend={error != null ? errorRender({ error }) : undefined}
-      cancel={
-        onCancel && (
-          <Button onClick={onCancel}>
-            {cancelLabel || <Trans>Cancel</Trans>}
-          </Button>
-        )
-      }
+      append={[append, error != null ? errorRender({ error }) : undefined]}
+      cancel={onCancel && <Button onClick={onCancel}>{cancelLabel}</Button>}
       actions={
-        <Button
-          color="primary"
-          type="submit"
-          loading={loading}
-          disabled={disabled}
-        >
-          {submitLabel || <Trans>Submit</Trans>}
-        </Button>
+        <>
+          <Button
+            color="primary"
+            type="submit"
+            loading={loading}
+            disabled={disabled}
+          >
+            {submitLabel}
+          </Button>
+          {actions}
+        </>
       }
     >
       {children}